Add files via upload
This code calculates the tip for a restaurant bill based on the total bill amount and the desired tip percentage. The tip is calculated by multiplying the bill amount by the tip percentage and dividing by 100.

Just add in my figures under the  bill amount and tip percentage. 

I will look to make this more complex  - split the bill between multiple people.  if someone can help me with this that would be great.